+++ warn/states/fl.py
@@ -6,32 +6,25 @@ from selenium.webdriver.common.by import By
 from selenium.common.exceptions import TimeoutException
 import pandas as pd
 
-from .base_warn import Warn
+from ..warn_base import Warn
 
-class GAWarn(Warn):
-    url = "https://www.tcsg.edu/warn-public-view/"
-    state = "GA"
+class FLWarn(Warn):
+    url = "https://reactwarn.floridajobs.org/WarnList/Records?year=2023"
+    state = "FL"
 
     def __init__(self, date=None):
         super().__init__(self.url, date)
-        self.tags = "#jobs #layoffs #GA #georgia"
+        self.tags = "#jobs #layoffs #FL #florida"
 
     def _fetch_latest_notices(self) -> dict:
         layoffs = {}
-        rows = self.get_html_rows()
-        if rows is None:
-            return layoffs
-        
-        df = rows[0]
+        df = self.get_html_rows()[0]
         if df is None:
             return {}
-        df["Submitted Date"] = pd.to_datetime(
-            df["Submitted Date"], 
-            format="%B %d, %Y",
-            errors="coerce"
-        )
-        df["Submitted Date"] = df["Submitted Date"].dt.strftime('%-m/%-d/%Y')
-        df = df[df["Submitted Date"] == self._compare_date]
+        
+        month, date, year = self.get_month_date_year(self._compare_date)
+        curr_date = f"{month}-{date}-{year}"
+        df = df[df["State Notification Date"] == curr_date]
         if len(df) == 0:
             return {}
         
@@ -39,7 +32,7 @@ class GAWarn(Warn):
             company_name = self.get_company_name(str(df["Company Name"]))
             if company_name is None:
                 continue
-            number_affected = row["Total Number of Affected Employees"]
+            number_affected = row["Employees Affected"]
             if company_name not in layoffs:
                 layoffs[company_name] = 0 
             layoffs[company_name] += number_affected
@@ -48,8 +41,8 @@ class GAWarn(Warn):
 
     def get_html_rows(self) -> list:
         self.driver.get(self._url)
-        delay = 5
-        xpath = "//th[@aria-label='Submitted Date: activate to sort column ascending']"
+        delay = 10
+        xpath = "//table[@id='DataTable']"
         try:
             WebDriverWait(self.driver, delay)\
                 .until(EC.presence_of_element_located((
@@ -66,7 +59,7 @@ class GAWarn(Warn):
             submitted_date_filter.click()
             WebDriverWait(self.driver, 2)
 
-            table_html = self.driver.find_element(By.ID, "DataTables_Table_0")\
+            table_html = self.driver.find_element(By.ID, "DataTable")\
                 .get_attribute('outerHTML')
             dfs = pd.read_html(table_html)
             return dfs
