+++ warn/states/ga.py
@@ -6,25 +6,32 @@ from selenium.webdriver.common.by import By
 from selenium.common.exceptions import TimeoutException
 import pandas as pd
 
-from .base_warn import Warn
+from ..warn_base import Warn
 
-class FLWarn(Warn):
-    url = "https://reactwarn.floridajobs.org/WarnList/Records?year=2023"
-    state = "FL"
+class GAWarn(Warn):
+    url = "https://www.tcsg.edu/warn-public-view/"
+    state = "GA"
 
     def __init__(self, date=None):
         super().__init__(self.url, date)
-        self.tags = "#jobs #layoffs #FL #florida"
+        self.tags = "#jobs #layoffs #GA #georgia"
 
     def _fetch_latest_notices(self) -> dict:
         layoffs = {}
-        df = self.get_html_rows()[0]
+        rows = self.get_html_rows()
+        if rows is None:
+            return layoffs
+        
+        df = rows[0]
         if df is None:
             return {}
-        
-        month, date, year = self.get_month_date_year(self._compare_date)
-        curr_date = f"{month}-{date}-{year}"
-        df = df[df["State Notification Date"] == curr_date]
+        df["Submitted Date"] = pd.to_datetime(
+            df["Submitted Date"], 
+            format="%B %d, %Y",
+            errors="coerce"
+        )
+        df["Submitted Date"] = df["Submitted Date"].dt.strftime('%-m/%-d/%Y')
+        df = df[df["Submitted Date"] == self._compare_date]
         if len(df) == 0:
             return {}
         
@@ -32,7 +39,7 @@ class FLWarn(Warn):
             company_name = self.get_company_name(str(df["Company Name"]))
             if company_name is None:
                 continue
-            number_affected = row["Employees Affected"]
+            number_affected = row["Total Number of Affected Employees"]
             if company_name not in layoffs:
                 layoffs[company_name] = 0 
             layoffs[company_name] += number_affected
@@ -41,8 +48,8 @@ class FLWarn(Warn):
 
     def get_html_rows(self) -> list:
         self.driver.get(self._url)
-        delay = 10
-        xpath = "//table[@id='DataTable']"
+        delay = 5
+        xpath = "//th[@aria-label='Submitted Date: activate to sort column ascending']"
         try:
             WebDriverWait(self.driver, delay)\
                 .until(EC.presence_of_element_located((
@@ -59,7 +66,7 @@ class FLWarn(Warn):
             submitted_date_filter.click()
             WebDriverWait(self.driver, 2)
 
-            table_html = self.driver.find_element(By.ID, "DataTable")\
+            table_html = self.driver.find_element(By.ID, "DataTables_Table_0")\
                 .get_attribute('outerHTML')
             dfs = pd.read_html(table_html)
             return dfs
