+++ warn/states/wa.py
@@ -2,43 +2,32 @@ from bs4 import BeautifulSoup
 import pandas as pd
 import requests
 
-from .base_warn import Warn
+from ..warn_base import Warn
 
-class OHWarn(Warn):
-    url = "https://jfs.ohio.gov/warn/current.stms"
-    state = "OH"
+class WAWarn(Warn):
+    url = "https://fortress.wa.gov/esd/file/warn/Public/SearchWARN.aspx"
+    state = "WA"
 
     def __init__(self, date=None):
         super().__init__(self.url, date)
-        self.tags = "#jobs #layoffs #OH #ohio"
+        self.tags = "#jobs #layoffs #WA #washington"
 
     def _fetch_latest_notices(self) -> dict:
         layoffs = {}
         df = self._get_html_table(self._url)
         if df is None:
             return {}
-
-        month, day, year = self._compare_date.split("/")
-        curr_date = "/".join([
-            month.zfill(2),
-            day.zfill(2),
-            year
-        ])
-
-        df = df[df["Date Received"] == curr_date]
+        
+        df = df[df["Received Date"] == self._compare_date]
         if len(df) == 0:
             return {}
 
         for _, row in df.iterrows():
             company_name = row["Company"]
-            number_affected = row["Potential Number Affected"]
+            number_affected = row["# of Workers"]
             if company_name not in layoffs:
                 layoffs[company_name] = 0 
-
-            try:
-                layoffs[company_name] += int(number_affected)
-            except:
-                layoffs[company_name] = number_affected
+            layoffs[company_name] += int(number_affected)
         return layoffs
   
     def _get_html_table(self, url:str=None) -> pd.DataFrame:
@@ -48,8 +37,10 @@ class OHWarn(Warn):
         try:
             response = requests.get(self._url)
             soup = BeautifulSoup(response.text, 'html.parser')
-            html_table = soup.find('table', {"class": "warnTable"})
-            dfs = pd.read_html(str(html_table))
-            return dfs[0]
+            html_table = soup.find('table', {"id": "ucPSW_gvMain"})
+            df = pd.read_html(str(html_table), match='Company')[0]
+            df = df.rename(columns=df.iloc[1])
+            df = df.iloc[2:]
+            return df
         except:
             return None
